Fix : Pop-up list writes value labels to model
In the pop-up list field property editor presentation, the logic for 
writing user selection to the model property uses the text of the list
entry. This could be a value label instead of the actual value if the
property has a ValueLabelService.

The fix uses the selection index to look up the original value from an
array constructed when the set of possible values is retrieved.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=386601