Fix : Pop-up list does not sort according to label
In the pop-up list field property editor presentation, the order of
items in the list is based on sorting by the actual value. If the
property has a ValueLabelService, the user sees value labels in the
list, so the perceived order may not be as expected. A better strategy
is to sort by value label.

The fix introduces PossibleValue class that holds both the value and the
label instead of using two parallel arrays. The new class implements
Comparable based on the label and the list of PossibleValue objects is
sorted prior to populating the pop-up list. This overrides the
value-based sort order produced by PossibleValuesService.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=386693