Fix : PropertyInitializationEvent is not issued
When property enablement is accessed, but property is not read, the
PropertyInitializationEvent is not issued. This can cause problems for
functionality dependent on that event.

For instance, @DependsOn (DependenciesService) feature uses
PropertyInitializationEvent to wire property refresh to its
dependencies. If the dependency declaration was added for the benefit of
an EnablementService, property enablement will not function correctly.

The fix ensures that full property refresh happens when
enablement of a property is accessed for the first time. This is an
overkill as property is read and validated when that's likely not
necessary. A better fix will need to be found.

The scenario is verified by TestModelingEvents0005.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=386715