Fix : Initial value gallery sample is incorrect
In the gallery sample, the section covering the initial value feature
was not updated after the fix for Bug 382453, which limited automatic
application of initial value to cases where the framework can determine
with certainty that the element is semantically new. An element is known
to be semantically new only when inserting a new element into a list or
creating a child element for a non-implicit element property.

The problem with the sample is that the property with @InitialValue is
located within two levels of implied element properties relative to the
model root. The initial value is not applied because at no point a
semantically explicit element creation is performed in the element
hierarchy containing this property.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=386862