Fix : Detail sections not displayed for node factory nodes
The conversion of visible when function from SapphireCondition to
Sapphire EL has introduced a regression in the master details editor
page. The detail sections are no longer shown for any node produced by
a node factory. Detail sections are still shown for nodes defined
statically.

The cause is an introduction of a node subclass to encapsulate logic
specific to nodes created by a node factory. This clashed with the code
that is responsible for serving up the detail sections. It is only
expecting one node class. The fix is to eliminate the node subclass and
make one class handle the case of a static node as well as the case of
a node factory node.