Fix : Facts not shown in property editor assist popup
When a property does not have content, but has a validation problem, the
user is able to access the property editor assist popup, but the popup
does not list the relevant facts about the property. The facts are only
listed if the property has content. Facts should be listed if property
has content or if property has a validation problem.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=387239