Enhancement : Overwrite existing file setting could be more obvious
The reusable form for creating workspace files validates for the
presence of an existing file with the specified name. To overwrite, user
must select the "overwrite existing file" option. This option is
unlikely to be missed as it is, but we could make this scenario slightly
easier by also presenting "overwrite existing file" as an automated
problem resolution. This will put both the problem and resolution in the
same popup.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=387248