Fix : Element without resource should initialize on instantiate
This change broke several unit tests that are verifying delivery 
of events. The issue is that initialize() reads implied element 
properties, which causes some events to be broadcast before returning 
from instantiate() call. 

The fix is to modify the affected tests to avoid calling the variant
of instantiate() that performs initialization.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=387332