Fix : Must be specified fact can be redundant
If a value property has a default value and @Required annotation, the 
following facts are generated:

Default value is "abc".
Must be specified.

The "must be specified" fact is redundant in presence of a default value
since the presence of a default value implies that @Required constraint
is always satisfied.

The fix includes test coverage in TestServices0001.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=387445