Enhancement : @PreferDefaultValue
In certain cases, deviations from property's default value are allowed,
but not recommended.

Consider a file creation wizard where the file should have a certain
name to function properly. This is common for configuration files. For
flexibility and consistency, the wizard should allow a different file
name to be used if the user so desires, but it should warn the user
that they are deviating from the file name recommendation.

The new @PreferDefaultValue annotation is used to specify the above
semantics. Utilizing this annotation produces a warning and a fact
statement.

Validation Warning

Format: [property] should be [default-value].
Example: File name should be "sapphire-extension.xml".

Fact

Format: Recommended value is [default-value].
Example: Recommended value is "sapphire-extension.xml".

This enhancement is verified by PreferDefaultValueTests.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=387498