Enhancement : FileName
When implementing operations that create files, a common need is to 
gather the name of a file from the user. The file name must be 
validated to be compatible with user's platform and certain 
normalization should be performed to simplify file name input.

This enhancement provides FileName class and associated services.

1. ValidationService – Flags invalid file names as errors.

2. ValueSerializationService – Only creates FileName objects that are 
   valid for the current platform.

3. ValueNormalizationService

   A. Leading whitespace is removed.
   
   B. Trailing whitespace and dots are removed.
   
   C. Extension is added if the file name does not have one already and 
      if the property has a FileExtensionsService (usually via 
      @FileExtensions annotation).
      
Also included in this enhancement are improvements to Path class
to perform validation of path segments analogous to validation
performed by FileName.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=387655