Enhancement : Merge ClassLocator and ResourceLocator
The existing ClassLocator and ResourceLocator abstractions enable
Sapphire to find classes and other resources typically loaded from
a class loader even if Sapphire is used in a context such as OSGi where
a ClassLoader instance may not always be exposed. The trouble is that
many situations require both of the locators and passing both around is
awkward. 

This enhancement merges ClassLocator and ResourceLocator into a single
abstraction:

Context
{
     static Context adapt( ClassLoader loader )
     static Context adapt( Class<?> cl )
     
     abstract <T> Class<T> findClass( String name )
     abstract URL findResource( String name )
     abstract List<URL> findResources( String name )
}

BundleBasedContext extends Context
{
     static Context adapt( Bundle bundle )
}

In addition to merging of the locators, this enhancement includes two
new features:

1. The findResources() method, which is used for locating all resources
   with the same name.

2. When Context is constructed based on a class instead of a class 
   loader, the implementation is capable of resolving classes and
   resources using a simple name relative to the package of context
   class.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=387741