Fix : Some services are accessed in the wrong context
Since adding parameter validation to ModelElement.services() methods,
a number of cases have been highlighted where a service for a child
element's property is requested from the parent element. These
situations arouse before there were separate Sapphire.Property.Instance
and Sapphire.Property.MetaModel service contexts.

The fix is to ensure that service implementations are registered in the
property meta-model context where possible. This ensures that various
functionality that needs to access the service is able to do so without
having an element instance. An example of such scenario is
a slush-bucket property editor presentation, where the set of possible
values for a property must be presented before any elements containing
this property are created.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=388035