Fix : Root XML element is not created
When an XML-based model is instantiated for a new file and subsequently
saved without any changes, the XML declaration and the root element are
not written to the file. This has led to the following workaround
applied prior to save:

( (XmlResource) element.resource() ).getXmlElement( true )

The workaround should not be necessary. In fact, since the workaround
is based on the file creator knowing that the model is XML-based, it
cannot be used in all situations. It would be better for RootXmlResource
on save to add the XML declaration and the root element if the document
is empty.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=388483