Enhancement : Instantiate a model without knowing resource type
In order to instantiate a model for a given input (such a file), one
needs to instantiate a particular type of a resource. The code that
needs to instantiate the model is not always in position to know the
type of a resource that is needed. There needs to be an API to
instantiate a model given an arbitrary input rather than a resource and
there needs to be a system for resource implementors to supply logic
for converting an input to a resource.

ModelElementType
{
    <T extends IModelElement> T instantiate( Resource resource ) [existing]
    <T extends IModelElement> T instantiate( Object input ) [added]
}

The new instantiate method variant is backed by new ConversionService
and MasterConversionService API. These services are more generic than
the described use case and will likely be used more broadly in the
future. In the model instantiation scenario, the ConversionService
implementations in the Sapphire.Element.MetaModel context are used.

ConversionService
{
    abstract <T> T convert( Object object, Class<T> type )
}

MasterConversionService has identical API to ConversionService. It
converts an object to the specified type by delegating to available
ConversionService implementations. If object is null or is already of
desired type, the object is returned unchanged. An implementation of
this service is provided with Sapphire. This service is not intended
to be implemented by adopters.

Two implementations of ConversionService are provided in the framework.

Sapphire.ConversionService.IFileToWorkspaceFileResourceStore 
  
  Capable of converting an IFile to a WorkspaceFileResourceStore or
  a ByteArrayResourceStore.

Sapphire.ConversionService.ByteArrayResourceStoreToXmlResource

  Capable of converting a ByteArrayResourceStore to an XmlResource or
  a Resource. Conversion is only performed if resource store corresponds
  to a file with "xml" extension or if the context element type has XML
  binding annotations.

A conversion service can delegate to other conversion services to create
a conversion chain. In fact, the two provided implementations are
commonly used together to convert an IFile to an XmlResource.

Two implementations of ConversionService are provided in the samples.

Calendar Sample

  A custom ConversionService is used due to the layered nature of the
  calendar model.

Purchase Order Sample

  A custom ConversionService is used because the default file extension
  for purchase order files is "po" rather than "xml" and PurchaseOrder
  element does not have XML binding annotations. The combination of
  these two factors prevent the framework-provided ConversionService
  implementation from engaging.

Putting all of this together, the following three snippets are
equivalent:

Resource resource = new RootXmlResource( new XmlResourceStore( new WorkspaceFileResourceStore( file ) ) );
PurchaseOrder po = PuchaseOrder.TYPE.instantiate( resource );

Resource resource = PurchaseOrder.TYPE.service( MasterConversionService.class ).convert( file, Resource.class );
PurchaseOrder po = PuchaseOrder.TYPE.instantiate( resource );

PurchaseOrder po = PuchaseOrder.TYPE.instantiate( file );

https://bugs.eclipse.org/bugs/show_bug.cgi?id=388485