Enhancement : Dynamic visibility for actions
The existing setEnabled() API makes it easy to dynamically control
enablement of an action, but dynamically controlling visibility is a lot
more difficult.

The current workaround is to wrap an action handler in a
SapphireActionHandlerFactory implementation. Since an action that has
no handlers is not visible and since a handler factory is allowed to
dynamically alter the set of handlers that it produces, this approach
allows one to dynamically control action visibility. This works, but it
is quite cumbersome.

Sapphire should support dynamic action visibility on par with how
dynamic action enablement is supported.

SapphireActionSystemPart
{
    boolean isEnabled() [existing]
    void setEnabled( boolean enabled ) [existing]
    boolean isVisible() [new]
    void setVisible( boolean visible ) [new]
}

https://bugs.eclipse.org/bugs/show_bug.cgi?id=389074