Continuation : Improved SWT and Eclipse RCP entry points
DefinitionLoader should allow a way to resolve a definition even if it
has validation errors. This is necessary when DefinitionLoader is used
during import from one definition to another. Refusing to load imported
definition is not a great solution in this case since the developer will
only be told that the imported definition could not be loaded instead
of the details of the underlying problem.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=388491