Fix : List property editor malfunctions in presence of @Required
The move up, move down and delete actions of a list property editor do
not behave as expected if the list entry property has an @Required
annotation.

1. Add @Required to IBrowseSupportGalleryListEntry.MultiOptionPath.

2. Open gallery sample and add a couple of entries in the first 
   MultiOptionPath editor. Make sure to use one of the browse handlers
   to specify the values for entries.
   
3. Notice that move up, move down and delete actions do not behave
   correctly.
   
The cause of this issue is an unnecessary list control refresh when list
property validation changes. The extra refresh is not just wasteful, but
is being processed in the wrong place, causing corruption to the list
property editor's state. 

The property editor should refresh the list control based on
PropertyContentEvent rather than PropertyEvent (super type).

The reason that @Required is exposing this issue is that it adds
validation.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=389013