Fix : Multiple instances of the same service are created
It is possible to cause two instances of the same service implementation
to be created in one context. This can lead to weird glitches if two
parts of a system attempt to communicate via this service.

It has been observed that given a service implementation S' that extends
service type S, if service lookups are made for S and S', then two
instances of S' will be created in the same service context.

ServiceContext maintains Service instances in a map keyed by class used
in the service lookup. This does not work. The solution is to hold
a flat list of service instances and to filter that list on lookup.

Test coverage in TestServices0012.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=387737