Fix : Glitches in pop-up list field property editor data binding
Several data binding glitches have been observed when exercising the
pop-up list field property editor presentation in the gallery sample
under Value Properties -> Possible Values.

The glitches range from not taking a change of value in certain cases
to not reacting to the restore default value command. 

It is possible that the glitches at least in part are due to having
multiple property editors with different presentations bound to the
same property instance.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=389635