Fix : Multiple XML unit test failures when launch from Eclipse
A number of failures in Sapphire XML support unit tests are reported
when launching tests from Eclipse. The failures are not seen when tests
are executed on the command line.

This issue has already been fixed in master branch as part of a larger
change, but the fix is also needed in R_0_5_x branch.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=389641