Fix : Inappropriate loading of extensions from another Sapphire version
In some scenarios (such as upgrade), multiple versions of Sapphire
bundles may be present and active. The Sapphire extension system is
not checking which version of Sapphire a particular bundle is associated
with before attempting to load its extensions. The result is a flood
of error reports regarding extension classes that don't appear to
extend the expected class. 

The fix is to load org.eclipse.sapphire.Sapphire class from a bundle
containing a Sapphire extension manifest. If the loaded class is not
the same as the one visible to the extension system, the extension
is skipped.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=389858