Enhancement : Collection factories should support filtering
Filtering is a common requirement when constructing a collection. The
collection factories should make filtering easier. It should be possible
apply a filter to a collection factory either before or after adding
items. When a filter is added, any elements previously added that are
not allowed by the filter should be removed. All elements added after
applying the filter should be run through the filter prior to adding.

ListFactory
{
    ListFactory<E> filter( Filter<E> filter ) [new]
}

MapFactory
{
    MapFactory<K,V> filter( Filter<Map.Entry<K,V>> filter ) [new]
}

SetFactory
{
    SetFactory<E> filter( Filter<E> filter ) [new]
}

SortedSetFactory
{
    SortedSetFactory<E> filter( Filter<E> filter ) [new]
}

https://bugs.eclipse.org/bugs/show_bug.cgi?id=389939