Fix : PossibleValuesService is inconsistent on case sensitivity
PossibleValuesService has isCaseSensitive() method which allows the
implementor to control case sensitivity. However, the SortedSet returned
by the values() method does not take this into account and is always
configured to be case-sensitive. This has lead to the need for consumers
of PossibleValuesService to deal with case-sensitivity on their own.

The SortedSet returned by the values() method should be configured with
a case-insensitive comparator, if appropriate. This should eliminate the
need for consumers of PossibleValuesService to ever reference the
isCaseSensitive() method.

An additional issue in this code is that no measure is taken to prevent
empty strings in the possible values set.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=390035