Enhancement : Better list of possible values story
Sapphire provides two property editors (a slush bucket and a check box
list) that cater to the list of possible values pattern. This pattern
is defined as a list property with a single entry type, which has a
single property, which happens to be a value property, and that value
property has a possible values constraint, and that value property does
not allow duplicates.

public interface Example extends IModelElement
{
    ModelElementType TYPE = new ModelElementType( Example.class );
    
    // *** FavoriteColors ***

    interface FavoriteColor extends IModelElement
    {
        ModelElementType TYPE = new ModelElementType( FavoriteColor.class );
        
        // *** Name ***
        
        @NoDuplicates
        @Service( impl = ColorsPossibleValuesService.class )

        ValueProperty PROP_NAME = new ValueProperty( TYPE, "Name" );
        
        Value<String> getName();
        void setName( String value );
    }
    
    @Type( base = MultiSelectStringItem.class )
    
    ListProperty PROP_FAVORITY_COLORS = new ListProperty( TYPE, "FavoriteColors" );
    
    ModelElementList<FavoriteColor> getMultiSelectString();
}

The implementation of the PossibleValueService may or may not need
access to the model in order to compute the set of possible values. This
isn't a problem for the standard table property editor. The only times
that this property editor interacts with the PossibleValuesService is to
show a pop-up list cell editor or to show a browse dialog for a text
field cell editor. In both of those cases, we are in the context of a
particular list entry, so we can access PossibleValuesService from the
property instance context. This gives the service implementation access
to the list entry and the rest of the model. The situation is more
difficult for the slush bucket and the check box list property editors.
Both of these property editors need to be able to retrieve the set of
possible values without a list entry instance. The current approach is
to access PossibleValuesService from property meta-model context, but
that approach only works for cases where the PossibleValuesService does
not need access to the model.

The solution is to allow a PossibleValuesService to be attached to the
list property as opposed to the value property in the list of possible
values scenario. The slush bucket and the check box list property editor
will now access the PossibleValuesService from the property instance
context of the list property, giving the implementation the ability to
access the model. To make this work, two new PossibleValueService
implementations are provided by the system. Both of these only
activate for the list of possible values pattern.

1. Maps PossibleValuesService from value property's meta-model context
   to list property's instance context. 
   
2. Maps PossibleValuesService from list property's instance context to
   value property's instance context.
   
The combination of the two new services allow the developer to provide
a PossibleValuesService implementation either at the value property's
meta-model context or at the list property's instance context, depending
on the need to access the model. The effect on model validation and the
behavior of the slush bucket and the check box property editors is
identical regardless of the chosen approach.

To demonstrate the new capability, this change modifies the gallery
sample's to allow custom colors to be defined, which can then be
selected in the Multi-Select section of the editor via a slush bucket
and a check box list editor samples.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=389763