Enhancement : Section description
A common pattern in forms involves a description text at the top of
a section to explain the contents of the section. This pattern can be
implemented using a label followed by a spacer in the section content,
but this approach fails to fully convey the developer's intent and can
lead to inconsistent presentation. A common mistake is failing to add
a spacer after the label.

Native support of section descriptions would be slightly easier to use
and avoid consistency issues.

<section>
    <label>shipping information</label>
    <description>If not specified, billing information is used for shipping purposes.</description>
    <content>
        ...
    </content>
</section>

https://bugs.eclipse.org/bugs/show_bug.cgi?id=390881