Enhancement : EL cast should attempt a cast through string
Sapphire EL performs type casts in order to convert the type of an
operand into the type that the operator expects. Any type can be cast
to a string and many types types can be cast from a string since that's
necessary in order to specify a literal. So, while a direct X-to-Y cast
may not be implemented, an indirect X-to-String-to-Y cast may be
possible.

If a direct cast is not possible, Sapphire EL should attempt an indirect
cast through a string before giving up.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=391609