Enhancement : SapphirePart.executeAfterInitialization()
In some cases, it is important to execute a piece of code after a part
has completed its initialization. PartInitializationEvent already
exists, but we need a way to check if a part has been initialized and
a convenience method to execute a job after a part has been initialized
or immediately if already initialized.

SapphirePart
{
    boolean initialized() [new]
    boolean executeAfterInitialization( Runnable job ) [new]
}

https://bugs.eclipse.org/bugs/show_bug.cgi?id=391621