Fix : Action does not enable
In certain situations, an action with multiple handlers does not enable
when its handlers become enabled.

The problem is limited to the case where an action handler factory is
used and this factory updates its list of handlers after the initial
set has been produced. This can happen with the add action when a
property has a custom PossibleTypesService.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=392672