Fix : DataService broadcasts too much on init
DataService performs an initial refresh during initialization. It is
possible for the data to be accessed prior to the initial refresh due
to re-entrant code paths. To cover for this, the refresh method
broadcasts a change even after the initial refresh. That's too noisy.
DataService should track if data has actually been accessed prior to
the initial refresh and use that information in deciding whether to
broadcast a change after the initial refresh.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=392856