Enhancement : Event tracing facility
Many of Sapphire performance problems are due to processing of
unnecessary events. To facility debugging such problems, we need a way
to trace events that are posted. The tracing facility should output
event details along with the source of the event (location in code from
where the event was posted). To control output verbosity, the tracing
facility should let the developer specify how many line of stack trace
to output when identifying the source of the event.

The tracing facility is controlled by a pair of statics:

boolean ListenerContext.TRACE
int Event.TRACE_SOURCE_STACK_DEPTH

The element and property events output all event details, while the
rest of the events only output event type and source. This aspect can
be enhanced in the future, as necessary.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=392937