Continuation : Versions and version constraints
Consider the case of an element property that is not empty, but whose
child properties are all empty. If this element property is presented
using "with" construct and becomes version incompatible, the user
will see the type selector, the line separator, and nothing beneath
the line separator. This is not visually appealing. A better
presentation would be to display a text message beneath the separator
informing the user that there are currently no properties that could be
displayed in that area. This is the same approach as what is used when
the element property is empty.

To illustrate this scenario, this change adds payment information to
the purchase order sample.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=383488