Fix : Duplicate connection added to diagram when creating connections
With the recent changes in event generation and broadcasting sequence 
in Sapphire model, our adopter product is adding duplicated connections
to the diagram when creating connections using connection creating tool.
This is caused by how model events are handled in 
DiagramConnectionTemplate class when it handles the addition of new 
element of the connection list property. It creates a new connection 
part and initializes it before recording the newly created connection 
part. The connection part initialization also triggers event which 
causes the event handler code to be re-executed. The fix is to record 
the newly created connection part before initializing it.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=393290