Fix : NPE thrown if the direct editing results the node being deleted
If a diagram connection is defined implicitly by relying on certain
pattern on the source node such as the following case:

        <control-flow-rule>
            <from-activity-id>foo*</from-activity-id>
            <control-flow-case>
                <from-outcome>success</from-outcome>
                <to-activity-id>methodCall0</to-activity-id>
            </control-flow-case>
        </control-flow-rule>

Here if the "from-activity-id" ends with "*", it'll be considered as a "wildcard" connection. When direct-editing the source node of the
connection, if user edit removes "*", then it's no longer considered a 
wildcard connection. Therefore the source node of the connection is 
removed from the diagram.

An NPE is thrown out of NodeDirectEditManager in bringDown() method
in such case. The fix is to get the graphical viewer from the diagram
editor.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=394334