Fix : Wizard banner image error if no wizard image specified
If no wizard banner image is specified in sdef at either wizard or page
level (can be seen in file bug report wizard of the EzBug sample), the
wizard displays a red square in the banner image area indicating an
error. The default wizard banner image should be used instead.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=394431