Fix : NPE thrown if the direct editing results the node being deleted
Part 2: when a node is removed by the direct-editing and if there are
connections attached to this node, it'd throw the following exception:

ava.lang.IllegalStateException: IControlFlowCase element is already disposed.
	at org.eclipse.sapphire.modeling.ModelElement.assertNotDisposed(ModelElement.java:1671)
	at org.eclipse.sapphire.modeling.ModelElement.attach(ModelElement.java:1223)
	at org.eclipse.sapphire.ui.diagram.editor.DiagramEmbeddedConnectionPart.addModelListener(DiagramEmbeddedConnectionPart.java:170)
	at org.eclipse.sapphire.ui.swt.gef.commands.LabelNodeCommand.execute(LabelNodeCommand.java:74)
	
The fix is to remove the connection parts that are attached to the node.
And in the direct edit command, before it adds the listeners back to 
connections, check whether the connection part has been disposed.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=394334