Fix : FailSafeFunction has a problem with malformed property values
FailSafeFunction implementation applies the default value before
performing the cast. This can lead to problems in certain situations
as it can be difficult to determine prior to the cast whether the value
is null. The answer can depend on which type is expected.

The current check is as follows:

if( val == null || ( val instanceof Value && ( (Value<?>) val ).getContent() == null ) )

The issue is with val.getContent(), which for a typed value property
will return null when property has a value, but that value is malformed.

The fix is to perform the cast prior to applying the default value. This
avoids dealing with the question of whether we should be referencing
the typed content or untyped text. The cast logic takes care of that.

This scenario is tested by TestExpr0018.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=397045