Fix : FailSafeFunction has a problem with malformed property values
The previous solution proved to be incorrect as it did not account for
the fact that EL cast to certain primitives changes null to that type's
inherent default value. For instance, casting a null to a boolean 
produces false. 

The revised solution unwraps Value objects in FailSafeFunction before
doing anything else. How unwrapping is done is dependent on which type
is expected from the fail-safe. If string is expected, value.getText()
is used, otherwise value.getContent() is used. This allows access to
the text of malformed values while preserving the expected behavior of
all defaults, including model-level defaults.

TestExpr0018 has been revised accordingly.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=397045