Releng : Change minimum requirement from Helios/3.6 to Indigo/3.7
Currently, Sapphire is compatible with all versions of Eclipse going
back to Helios/3.6 release. As part of improvements to the diagram
editor functionality, we are running into GEF limitations in Helios/3.6
release that would be cumbersome to workaround. Moving the minimum
requirement to Indigo/3.7 allows us to use GEF API introduced in that
release.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=397724