Enhancement : Service registration without a factory
When services are registered through Sapphire extension system, the
service implementation class is not specified directly. Instead,
a ServiceFactory is implemented, which enables applicability of the
service to be checked. For many services, specifying a service context
is sufficient and a custom applicability check is not necessary. To make
it easier to register such services, the extension system should support
registering services without using a factory.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=398437