Enhancement : Control context of services registered via annotations
Currently, services registered via annotations on elements and
properties are placed in both metamodel and instance contexts. This is
rarely the correct behavior. More commonly, the service should be in one
context or the other. In many cases, a service cannot function in the
metamodel context because it needs access to the element instance.
The @Service annotation should have an attribute to pick between
metamodel and instance contexts. The instance context should be the
default.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=398443