Enhancement : Extensible persistent state for editor pages
Editor pages are able to persist user interface state between sessions
independent of the data that is being edited. What state is persisted is
dependent on editor page type. Two common examples of persistent state
are sizing of resizable elements and selections.

Adopters should be able to extend the persistent state to persist
additional data for custom actions or parts.

Two extensibility approaches are supported:

1. Extend the page's persistent state element and specify the custom
   element type in sdef.

2. Store custom state as arbitrary key-value pairs without extending
   the persistent state element. This feature integrates with Sapphire's
   library of conversions to make it typically not necessary to manually
   convert attribute values to and from string.

Both approaches are demonstrated in the catalog sample. See the
Show Manufacturer and Color Code Manufacturers toggles in the editor
page header.

This feature is documented in the developer guide under "Extensible
Persistent State" heading.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=392842