Fix : Slush bucket property editor should not always suppress add action
In most cases, slush bucket property editor works with properties where
deviation from possible values is an error. In these cases, it is
appropriate to hide the standard add action as the left side of the
slush bucket lists all possibilities.

When deviation from possible values is not an error, the slush bucket
property editor should show the standard add action in order to
facilitated arbitrary item entry.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=399784