Enhancement : Class for Java identifiers
In models that reference Java resources, it is sometimes necessary to
reference Java identifiers. An identifier is a name of a variable, a
field or a method. Identifiers must conform to [a-zA-Z_$][a-zA-Z0-9_$]*
pattern.

The JavaIdentifier class performs verification in the constructor, so
any instance can be assumed to represent a valid identifier. This class
can be used by itself or as a type of a value property.

Test coverage in ConversionTests.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=400279