Enhancement : EL support in @Required
In certain scenarios, the required constraint is not static. It should
be possible to use EL in the @Required annotation or to supply a custom
service. 

This change introduces RequiredConstraintService. The @Required
annotation is now handled by DeclarativeRequiredConstraintService,
which support EL in @Required. RequiredConstraintValidationService and
RequiredConstraintFactsService now work through
RequiredConstraintService.

Test coverage provided by RequiredConstraintTests.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=376330