Fix : Many EL operators do not work with two properties
Many of the EL operators (such as comparison operators and arithmetic
operators) fail when used with two value properties. The operators work
correctly when applied to one property and one literal. The issue is
that values need to be unboxed before performing type analysis employed
in these operators.

Test coverage provided by OperatorTests.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=400868