Enhancement : Support ordered possible values
Typically, possible values have no particular order and it is desirable
to present them sorted. In some cases, the order is significant.

An example of ordered possible values is task severity. The severities
need to be presented in the severity rank order, not in the alphabetical
order.

There needs to be a way to indicate that possible values are ordered.

@PossibleValues
{
    ...
    boolean ordered() default false
}

PossibleValuesService
{
    ...
    boolean ordered()
}

The following presentations respect the new directive:

1. Slush bucket property editor
2. Checkbox list property editor
3. Pop-up list field property editor

The possible values browse dialog does not respect the new directive.
This is because the dialog is based on an existing filter dialog that
always sorts and addressing this would require forking the filter
dialog.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=401956