Enhancement : Flexible diagram node shapes
In Sample Schema editor, "direct edit" context menu is displayed when
column is selected, but is not displayed when the table node is
selected. We should not have this context menu. The fix is to introduce
another type of action context for a shape.

Bugzilla : https://bugs.eclipse.org/bugs/show_bug.cgi?id=404503