Enhancement : Flexible diagram node shapes
Fix the index of the figure of a child shape within the container shape
figure. The index is used to add a shape's figure to the container
figure when the shape becomes visible. The figure needs to be added at
the right place in order to maintain valid sequence layout.

Bugzilla : https://bugs.eclipse.org/bugs/show_bug.cgi?id=383924