Enhancement : Simpler way to specify date serialization
When working with date values, frequently the formats must be specified
explicitly instead of relying on the default formats. Currently, this
is done by parameterizing or subclassing the default conversion service
implementations. This is an awkward and verbose approach. A better
approach would be to use an annotation to supply formats for a given
property instance.

@Serialization( primary = "yyyy-MM-dd", alternative = "MM/dd/yyyy" )

Bugzilla : https://bugs.eclipse.org/bugs/show_bug.cgi?id=411540